fix: 챌린지 화면 리뷰 피드백 반영

+++ app/src/main/java/com/example/onuldo_fe/ui/screen/challenge/detail/DetailRoute.kt
@@ -57,7 +57,7 @@ fun DetailRoute(
     if (detail != null) {
         DetailScreen(
             challenge = Challenge(
-                id = detail.id.toInt(),
+                id = detail.id,
                 title = detail.title,
                 participantCount = detail.participantCount,
                 category = detail.category
